Update dependencies and toolchain
* rust nightly-2022-06-14 -> nightly-2024-02-24
* bitflags 1.2 -> 2.4
* iced 589f68 -> 0.12
* iced_native replaced with iced_runtime 0.12
* iced_web 589f68 -> 0.4
* wgpu 0.10.1 -> 0.13
* pixels 0.6.0 -> 0.13
* winit 0.25.0 -> 0.29

+++ gb_core/tests/ppu_monochrome.rs
@@ -37,7 +37,7 @@ fn ppu_singlecolor() {
         frame.iter().for_each(|&pix| {
             assert_eq!(
                 pix,
-                color::COLORS[color::calculate_monochrome_color_id(ppu.bgp, color) as usize]
+                color::COLORS[color::calculate_monochrome_color_id(ppu.bgp, color)]
             )
         });
     }
